added booking forms and moved auth forms
